# artifacts/api-server/src/routes/qr.ts
import { Router, type IRouter, type Request, type Response } from "express";
import QRCode from "qrcode";
import sharp from "sharp";
import { v4 as uuidv4 } from "uuid";
import { db, qrSessionsTable, visitsTable } from "@workspace/db";
import { eq, desc } from "drizzle-orm";

const router: IRouter = Router();

interface TextOverlayConfig {
  text: string;
  font: string;
  size: number;
  x: number;
  y: number;
  color: string;
}

interface QrGenerateRequest {
  qrData: string;
  dynamicContent?: string;
  cornerColor?: string;
  logoUrl?: string;
  logoScale?: number;
  textOverlay?: TextOverlayConfig;
  whiteFill?: boolean;
  whiteFillColor?: string;
  highlightCorners?: boolean;
  metadataFriendly?: boolean;
  pageTitle?: string;
  pageDescription?: string;
  vocalIntroUrl?: string;
  trackTraffic?: boolean;
  tabsToOpen?: number;
  premiumFeatures?: boolean;
}

async function hexToRgb(hex: string): Promise<{ r: number; g: number; b: number }> {
  const clean = hex.replace("#", "");
  const r = parseInt(clean.substring(0, 2), 16);
  const g = parseInt(clean.substring(2, 4), 16);
  const b = parseInt(clean.substring(4, 6), 16);
  return { r, g, b };
}

async function generateQrWithOptions(
  data: string,
  opts: QrGenerateRequest
): Promise<string> {
  const cornerColor = opts.cornerColor || "#000000";
  const highlightCorners = opts.highlightCorners !== false;
  const logoScale = Math.min(opts.logoScale ?? 0.2, 0.4);

  const darkColor = highlightCorners ? cornerColor : "#000000";

  const qrSize = 512;
  const qrBuffer = await QRCode.toBuffer(data, {
    errorCorrectionLevel: "H",
    type: "png",
    width: qrSize,
    margin: 2,
    color: {
      dark: darkColor,
      light: opts.whiteFill ? (opts.whiteFillColor || "#FAFAFA") : "#FFFFFF",
    },
  });

  let image = sharp(qrBuffer).resize(qrSize, qrSize);

  const composites: sharp.OverlayOptions[] = [];

  if (opts.logoUrl && logoScale > 0) {
    try {
      const logoSize = Math.floor(qrSize * logoScale);
      const logoX = Math.floor((qrSize - logoSize) / 2);
      const logoY = Math.floor((qrSize - logoSize) / 2);

      const fetch = (await import("node:https")).get;
      const logoBuffer = await new Promise<Buffer>((resolve, reject) => {
        const protocol = opts.logoUrl!.startsWith("https") ? require("https") : require("http");
        protocol.get(opts.logoUrl!, (res: any) => {
          const chunks: Buffer[] = [];
          res.on("data", (chunk: Buffer) => chunks.push(chunk));
          res.on("end", () => resolve(Buffer.concat(chunks)));
          res.on("error", reject);
        }).on("error", reject);
      });

      const resizedLogo = await sharp(logoBuffer)
        .resize(logoSize, logoSize, { fit: "contain", background: { r: 255, g: 255, b: 255, alpha: 0 } })
        .toBuffer();

      composites.push({ input: resizedLogo, left: logoX, top: logoY });
    } catch (_e) {
    }
  }

  if (opts.textOverlay && opts.textOverlay.text) {
    const ov = opts.textOverlay;
    const x = Math.floor((ov.x / 100) * qrSize);
    const y = Math.floor((ov.y / 100) * qrSize);
    const fontSize = Math.max(8, Math.min(ov.size, 120));

    const fontWeight = ov.font === "bold" ? "bold" : "normal";
    const fontStyle = ov.font === "italic" ? "italic" : "normal";
    const fontFamily = ov.font === "cursive"
      ? "Georgia, serif"
      : "Arial, sans-serif";

    const svgText = `
      <svg width="${qrSize}" height="${qrSize}" xmlns="http://www.w3.org/2000/svg">
        <text
          x="${x}"
          y="${y}"
          font-family="${fontFamily}"
          font-size="${fontSize}"
          font-weight="${fontWeight}"
          font-style="${fontStyle}"
          fill="${ov.color || "#000000"}"
          dominant-baseline="middle"
          text-anchor="middle"
        >${ov.text.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;")}</text>
      </svg>`;

    composites.push({ input: Buffer.from(svgText), top: 0, left: 0 });
  }

  if (composites.length > 0) {
    image = sharp(await image.toBuffer()).composite(composites);
  }

  const finalBuffer = await image.png().toBuffer();
  return finalBuffer.toString("base64");
}

function buildPageHtml(session: typeof qrSessionsTable.$inferSelect): string {
  const title = session.pageTitle || "QR Code Page";
  const description = session.pageDescription || "Generated by ACB Dynamic QR Studio";
  const data = session.dynamicContent || session.qrData;

  const metaTags = session.metadataFriendly
    ? `
    <meta name="description" content="${description.replace(/"/g, "&quot;")}">
    <meta property="og:title" content="${title.replace(/"/g, "&quot;")}">
    <meta property="og:description" content="${description.replace(/"/g, "&quot;")}">
    <meta name="robots" content="index, follow">
    `
    : "";

  const audioTag = session.premiumFeatures && session.vocalIntroUrl
    ? `<audio autoplay src="${session.vocalIntroUrl}"></audio>`
    : "";

  const trackingScript = session.premiumFeatures && session.trackTraffic
    ? `<script>
        fetch('/api/traffic/${session.id}/track', {
          method: 'POST',
          headers: {'Content-Type': 'application/json'},
          body: JSON.stringify({ source: document.referrer || 'direct', userAgent: navigator.userAgent })
        }).catch(() => {});
       </script>`
    : "";

  return `<!DOCTYPE html>
<html lang="en">
<head>
  <meta charset="UTF-8">
  <meta name="viewport" content="width=device-width, initial-scale=1.0">
  <title>${title.replace(/</g, "&lt;")}</title>
  ${metaTags}
</head>
<body style="margin:0;display:flex;flex-direction:column;align-items:center;justify-content:center;min-height:100vh;background:#0f0f0f;font-family:sans-serif;color:#fff;">
  ${audioTag}
  <h1 style="margin-bottom:1rem;font-size:1.5rem;">${title.replace(/</g, "&lt;")}</h1>
  <img
    src="data:image/png;base64,${session.qrImageBase64}"
    alt="${description.replace(/"/g, "&quot;")}"
    style="max-width:400px;width:90%;border-radius:12px;box-shadow:0 8px 32px rgba(0,0,0,.5);"
    width="400"
    height="400"
  />
  <p style="margin-top:1rem;color:#aaa;font-size:0.875rem;">${data.replace(/</g, "&lt;")}</p>
  ${trackingScript}
</body>
</html>`;
}

router.post("/generate", async (req: Request, res: Response) => {
  try {
    const body = req.body as QrGenerateRequest;

    if (!body.qrData) {
      res.status(400).json({ error: "MISSING_QR_DATA", message: "qrData is required" });
      return;
    }

    const effectiveData = body.dynamicContent?.trim() || body.qrData;
    const id = uuidv4();

    const qrImageBase64 = await generateQrWithOptions(effectiveData, body);
    const pageUrl = `/api/qr/page/${id}`;

    const session = {
      id,
      qrData: body.qrData,
      dynamicContent: body.dynamicContent || null,
      cornerColor: body.cornerColor || "#000000",
      logoUrl: body.logoUrl || null,
      logoScale: body.logoScale ?? 0.2,
      textOverlay: body.textOverlay ? JSON.parse(JSON.stringify(body.textOverlay)) : null,
      whiteFill: body.whiteFill ?? false,
      whiteFillColor: body.whiteFillColor || "#FFFFFF",
      highlightCorners: body.highlightCorners !== false,
      metadataFriendly: body.metadataFriendly !== false,
      pageTitle: body.pageTitle || null,
      pageDescription: body.pageDescription || null,
      vocalIntroUrl: body.vocalIntroUrl || null,
      trackTraffic: body.trackTraffic ?? false,
      tabsToOpen: body.tabsToOpen ?? 1,
      premiumFeatures: body.premiumFeatures ?? false,
      qrImageBase64,
      pageUrl,
    };

    const [inserted] = await db.insert(qrSessionsTable).values(session).returning();

    res.json(inserted);
  } catch (err) {
    req.log.error(err, "Failed to generate QR");
    res.status(500).json({ error: "GENERATION_FAILED", message: "QR generation failed" });
  }
});

router.get("/sessions", async (req: Request, res: Response) => {
  try {
    const sessions = await db
      .select()
      .from(qrSessionsTable)
      .orderBy(desc(qrSessionsTable.createdAt));
    res.json(sessions);
  } catch (err) {
    req.log.error(err, "Failed to list sessions");
    res.status(500).json({ error: "DB_ERROR", message: "Failed to fetch sessions" });
  }
});

router.get("/sessions/:id", async (req: Request, res: Response) => {
  try {
    const [session] = await db
      .select()
      .from(qrSessionsTable)
      .where(eq(qrSessionsTable.id, req.params.id));

    if (!session) {
      res.status(404).json({ error: "NOT_FOUND", message: "Session not found" });
      return;
    }
    res.json(session);
  } catch (err) {
    req.log.error(err, "Failed to get session");
    res.status(500).json({ error: "DB_ERROR", message: "Failed to fetch session" });
  }
});

router.delete("/sessions/:id", async (req: Request, res: Response) => {
  try {
    await db.delete(qrSessionsTable).where(eq(qrSessionsTable.id, req.params.id));
    await db.delete(visitsTable).where(eq(visitsTable.sessionId, req.params.id));
    res.json({ success: true, id: req.params.id });
  } catch (err) {
    req.log.error(err, "Failed to delete session");
    res.status(500).json({ error: "DB_ERROR", message: "Failed to delete session" });
  }
});

router.post("/test", async (req: Request, res: Response) => {
  try {
    const { sessionId, qrImageBase64 } = req.body as { sessionId?: string; qrImageBase64?: string };

    let session: typeof qrSessionsTable.$inferSelect | undefined;

    if (sessionId) {
      const [found] = await db
        .select()
        .from(qrSessionsTable)
        .where(eq(qrSessionsTable.id, sessionId));
      session = found;
    }

    const issues: string[] = [];
    const warnings: string[] = [];
    const recommendations: string[] = [];
    let logoConflict = false;
    let cornerConflict = false;
    let confidence = 0.95;

    if (session) {
      const logoScale = session.logoScale ?? 0;
      if (logoScale > 0.3) {
        logoConflict = true;
        confidence -= 0.25;
        issues.push("Logo scale exceeds 30% — critical QR data modules may be obscured");
        recommendations.push("Reduce logo scale to 0.25 or below for reliable scanning");
      } else if (logoScale > 0.25) {
        warnings.push("Logo scale is above 25% — some scanners may struggle");
        recommendations.push("Consider reducing logo scale to 0.2 for best compatibility");
        confidence -= 0.1;
      }

      if (session.whiteFill && session.whiteFillColor !== "#FFFFFF") {
        warnings.push("Custom white fill color may reduce QR contrast");
        recommendations.push("Use high-contrast colors for fill to maintain scannability");
        confidence -= 0.05;
      }

      if (session.highlightCorners && session.cornerColor !== "#000000") {
        const hex = session.cornerColor.replace("#", "");
        const r = parseInt(hex.substring(0, 2), 16);
        const g = parseInt(hex.substring(2, 4), 16);
        const b = parseInt(hex.substring(4, 6), 16);
        const luminance = (0.299 * r + 0.587 * g + 0.114 * b) / 255;
        if (luminance > 0.6) {
          cornerConflict = true;
          confidence -= 0.2;
          issues.push("Corner color is too light — finder squares may not be detected");
          recommendations.push("Use a darker corner color (luminance below 40%) for finder square visibility");
        } else if (luminance > 0.4) {
          warnings.push("Corner color has moderate luminance — test with multiple scanners");
          confidence -= 0.08;
        }
      }

      if (session.qrData && session.qrData.length > 200) {
        warnings.push("QR data is long — use a URL shortener for better reliability");
        confidence -= 0.05;
      }

      if (session.textOverlay) {
        warnings.push("Text overlay may cover QR modules — verify scannability after placement");
        confidence -= 0.05;
      }
    } else {
      warnings.push("No session found — running generic validation only");
      recommendations.push("Generate a QR session first for detailed analysis");
    }

    if (issues.length === 0 && warnings.length === 0) {
      recommendations.push("QR code looks great! No issues detected.");
    }

    confidence = Math.max(0, Math.min(1, confidence));

    res.json({
      sessionId: session?.id,
      scannable: confidence >= 0.6,
      confidence,
      issues,
      warnings,
      recommendations,
      logoConflict,
      cornerConflict,
    });
  } catch (err) {
    req.log.error(err, "Failed to test QR");
    res.status(500).json({ error: "TEST_FAILED", message: "QR test failed" });
  }
});

router.get("/page/:id", async (req: Request, res: Response) => {
  try {
    const [session] = await db
      .select()
      .from(qrSessionsTable)
      .where(eq(qrSessionsTable.id, req.params.id));

    if (!session) {
      res.status(404).json({ error: "NOT_FOUND", message: "Session not found" });
      return;
    }

    const contentType = req.headers.accept?.includes("text/html") ? "html" : "json";

    if (contentType === "html") {
      res.setHeader("Content-Type", "text/html; charset=utf-8");
      res.send(buildPageHtml(session));
    } else {
      res.json({
        sessionId: session.id,
        html: buildPageHtml(session),
        pageTitle: session.pageTitle,
        pageDescription: session.pageDescription,
      });
    }
  } catch (err) {
    req.log.error(err, "Failed to get QR page");
    res.status(500).json({ error: "DB_ERROR", message: "Failed to fetch page" });
  }
});

export default router;