# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ACB Dynamic QR Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
  message: string;
}

export interface DeleteResult {
  success: boolean;
  id: string;
}

export type TextOverlayConfigFont =
  (typeof TextOverlayConfigFont)[keyof typeof TextOverlayConfigFont];

export const TextOverlayConfigFont = {
  bold: "bold",
  italic: "italic",
  cursive: "cursive",
  normal: "normal",
} as const;

export interface TextOverlayConfig {
  text: string;
  font: TextOverlayConfigFont;
  /**
   * @minimum 8
   * @maximum 200
   */
  size: number;
  /** X position as percentage of image width (0-100) */
  x: number;
  /** Y position as percentage of image height (0-100) */
  y: number;
  /** CSS color string for text */
  color: string;
}

export interface QrGenerateRequest {
  /** Primary data to encode (URL, text, etc.) */
  qrData: string;
  /** Runtime override for qrData; if present, used instead */
  dynamicContent?: string;
  /** Hex color for corner finder squares (e.g. "#FF0000") */
  cornerColor?: string;
  /** URL of the central logo image */
  logoUrl?: string;
  /**
   * Logo size as fraction of QR width (max 0.4 to preserve scannability)
   * @minimum 0
   * @maximum 0.4
   */
  logoScale?: number;
  textOverlay?: TextOverlayConfig;
  /** Apply paint-bucket fill to white QR modules */
  whiteFill?: boolean;
  /** Color to fill white areas with (default transparent/off-white) */
  whiteFillColor?: string;
  /** Apply cornerColor to the QR finder squares */
  highlightCorners?: boolean;
  /** Include SEO meta tags and alt text in generated page */
  metadataFriendly?: boolean;
  /** HTML page title for the generated page */
  pageTitle?: string;
  /** Meta description for the generated page */
  pageDescription?: string;
  /** URL of audio file to autoplay on page load (premium) */
  vocalIntroUrl?: string;
  /** Enable traffic tracking for this session (premium) */
  trackTraffic?: boolean;
  /**
   * Number of browser tabs this session targets
   * @minimum 1
   * @maximum 10
   */
  tabsToOpen?: number;
  /** Enable premium features (traffic tracking, vocal intro) */
  premiumFeatures?: boolean;
}

export interface QrSession {
  id: string;
  qrData: string;
  dynamicContent?: string;
  cornerColor?: string;
  logoUrl?: string;
  logoScale?: number;
  textOverlay?: TextOverlayConfig;
  whiteFill?: boolean;
  whiteFillColor?: string;
  highlightCorners?: boolean;
  metadataFriendly?: boolean;
  pageTitle?: string;
  pageDescription?: string;
  vocalIntroUrl?: string;
  trackTraffic?: boolean;
  tabsToOpen?: number;
  premiumFeatures?: boolean;
  /** Base64-encoded PNG of the generated QR code */
  qrImageBase64: string;
  /** URL path to the rendered HTML page for this session */
  pageUrl: string;
  createdAt: string;
  updatedAt: string;
}

export interface QrTestRequest {
  /** ID of the session to test */
  sessionId?: string;
  /** Raw base64 QR image to test (alternative to sessionId) */
  qrImageBase64?: string;
}

export interface QrTestResult {
  sessionId?: string;
  /** Whether the QR code is reliably scannable */
  scannable: boolean;
  /**
   * Scan reliability score (0-1)
   * @minimum 0
   * @maximum 1
   */
  confidence: number;
  /** List of detected issues */
  issues: string[];
  /** List of warnings that may affect scannability */
  warnings: string[];
  /** Suggestions for improving the QR code */
  recommendations: string[];
  /** Whether the logo overlaps critical QR data modules */
  logoConflict: boolean;
  /** Whether corner highlighting may reduce reliability */
  cornerConflict: boolean;
}

export interface QrPage {
  sessionId: string;
  /** Full HTML string for the generated page */
  html: string;
  pageTitle?: string;
  pageDescription?: string;
}

/**
 * Map of referrer/source to visit count
 */
export type TrafficReportSources = { [key: string]: number };

export interface Visit {
  id: string;
  sessionId: string;
  source?: string;
  userAgent?: string;
  visitedAt: string;
}

export interface TrafficReport {
  sessionId: string;
  totalVisits: number;
  uniqueVisitors: number;
  /** Map of referrer/source to visit count */
  sources: TrafficReportSources;
  recentVisits: Visit[];
}

export interface TrackVisitRequest {
  /** Referrer or traffic source label */
  source?: string;
  userAgent?: string;
}

export interface TrackVisitResult {
  success: boolean;
  visitId: string;
}