# lib/api-zod/src/generated/types/deleteResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ACB Dynamic QR Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface DeleteResult {
  success: boolean;
  id: string;
}