# lib/api-zod/src/generated/types/qrGenerateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ACB Dynamic QR Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { TextOverlayConfig } from "./textOverlayConfig";

export interface QrGenerateRequest {
  /** Primary data to encode (URL, text, etc.) */
  qrData: string;
  /** Runtime override for qrData; if present, used instead */
  dynamicContent?: string;
  /** Hex color for corner finder squares (e.g. "#FF0000") */
  cornerColor?: string;
  /** URL of the central logo image */
  logoUrl?: string;
  /**
   * Logo size as fraction of QR width (max 0.4 to preserve scannability)
   * @minimum 0
   * @maximum 0.4
   */
  logoScale?: number;
  textOverlay?: TextOverlayConfig;
  /** Apply paint-bucket fill to white QR modules */
  whiteFill?: boolean;
  /** Color to fill white areas with (default transparent/off-white) */
  whiteFillColor?: string;
  /** Apply cornerColor to the QR finder squares */
  highlightCorners?: boolean;
  /** Include SEO meta tags and alt text in generated page */
  metadataFriendly?: boolean;
  /** HTML page title for the generated page */
  pageTitle?: string;
  /** Meta description for the generated page */
  pageDescription?: string;
  /** URL of audio file to autoplay on page load (premium) */
  vocalIntroUrl?: string;
  /** Enable traffic tracking for this session (premium) */
  trackTraffic?: boolean;
  /**
   * Number of browser tabs this session targets
   * @minimum 1
   * @maximum 10
   */
  tabsToOpen?: number;
  /** Enable premium features (traffic tracking, vocal intro) */
  premiumFeatures?: boolean;
}